+++ common/src/main/java/io/netty/util/internal/jzlib/Tree.java
@@ -213,7 +213,7 @@ final class Tree {
                     continue;
                 }
                 if (tree[m * 2 + 1] != bits) {
-                    s.opt_len += ((long) bits - (long) tree[m * 2 + 1]) *
+                    s.opt_len += ((long) bits - tree[m * 2 + 1]) *
                             tree[m * 2];
                     tree[m * 2 + 1] = (short) bits;
                 }
